Move the Page Interface around
